+++ LinearizedTrackFit/interface/SingleSector.h
@@ -11,7 +11,7 @@
 namespace LinearFit {
 
   void singleSector(const TString & inputFileName, const double & eventsFractionStart, const double & eventsFractionEnd,
-      const std::unordered_map<std::string, std::unordered_set<int> > & requiredLayers, std::vector<float> distanceCuts,
+      const std::unordered_map<std::string, std::unordered_set<int> > & requiredLayers, std::vector<double> distanceCuts,
       const std::vector<std::string> & inputVarNames, const std::vector<std::string> & inputTrackParameterNames)
   {
     TreeReader treeReader(inputFileName, eventsFractionStart, eventsFractionEnd, requiredLayers, distanceCuts, inputVarNames, inputTrackParameterNames);
