+++ LinearizedTrackFit/interface/TreeReader.h
@@ -19,7 +19,7 @@ class TreeReader
 {
 public:
   TreeReader(const TString & inputFileName, const double & eventsFractionStart, const double & eventsFractionEnd,
-      const std::unordered_map<std::string, std::unordered_set<int> > & requiredLayers, const std::vector<float> & distanceCuts,
+      const std::unordered_map<std::string, std::unordered_set<int> > & requiredLayers, const std::vector<double> & distanceCuts,
       const std::vector<std::string> & varNames, const std::vector<std::string> & trackParNames);
 
   bool nextTrack();
@@ -76,7 +76,7 @@ private:
   std::vector<float> parameters_;
   unsigned int maxRequiredLayers_;
   unsigned int variablesSize_;
-  std::vector<float> distanceCuts_;
+  std::vector<double> distanceCuts_;
 
   std::vector<StubRZPhi> stubsRZPhi_;
 };
